🚚 refactor: Changed langchain module directory to inside the material module directory.
