+++ src/modules/material/material.module.ts
@@ -5,7 +5,7 @@ import { AuthModule } from '@modules/auth/auth.module';
 import { UploadModule } from '@modules/upload/upload.module';
 import { PrismaModule } from '@modules/prisma/prisma.module';
 import { MaterialRepository } from './material.repository';
-import { LangchainModule } from '@modules/langchain/langchain.module';
+import { LangchainModule } from '@modules/material/langchain/langchain.module';
 
 @Module({
   imports: [
